Move updater and downloader to web module
