+++ src/main/java/org/betonquest/betonquest/BetonQuest.java
@@ -149,15 +149,15 @@ import org.betonquest.betonquest.modules.schedule.impl.realtime.cron.RealtimeCro
 import org.betonquest.betonquest.modules.schedule.impl.realtime.cron.RealtimeCronScheduler;
 import org.betonquest.betonquest.modules.schedule.impl.realtime.daily.RealtimeDailySchedule;
 import org.betonquest.betonquest.modules.schedule.impl.realtime.daily.RealtimeDailyScheduler;
-import org.betonquest.betonquest.modules.updater.UpdateDownloader;
-import org.betonquest.betonquest.modules.updater.UpdateSourceHandler;
-import org.betonquest.betonquest.modules.updater.Updater;
-import org.betonquest.betonquest.modules.updater.source.DevelopmentUpdateSource;
-import org.betonquest.betonquest.modules.updater.source.ReleaseUpdateSource;
-import org.betonquest.betonquest.modules.updater.source.implementations.GitHubReleaseSource;
-import org.betonquest.betonquest.modules.updater.source.implementations.NexusReleaseAndDevelopmentSource;
 import org.betonquest.betonquest.modules.versioning.Version;
 import org.betonquest.betonquest.modules.versioning.java.JREVersionPrinter;
+import org.betonquest.betonquest.modules.web.updater.UpdateDownloader;
+import org.betonquest.betonquest.modules.web.updater.UpdateSourceHandler;
+import org.betonquest.betonquest.modules.web.updater.Updater;
+import org.betonquest.betonquest.modules.web.updater.source.DevelopmentUpdateSource;
+import org.betonquest.betonquest.modules.web.updater.source.ReleaseUpdateSource;
+import org.betonquest.betonquest.modules.web.updater.source.implementations.GitHubReleaseSource;
+import org.betonquest.betonquest.modules.web.updater.source.implementations.NexusReleaseAndDevelopmentSource;
 import org.betonquest.betonquest.notify.ActionBarNotifyIO;
 import org.betonquest.betonquest.notify.AdvancementNotifyIO;
 import org.betonquest.betonquest.notify.BossBarNotifyIO;
