+++ src/main/java/org/betonquest/betonquest/commands/QuestCommand.java
@@ -32,13 +32,13 @@ import org.betonquest.betonquest.id.EventID;
 import org.betonquest.betonquest.id.ItemID;
 import org.betonquest.betonquest.id.ObjectiveID;
 import org.betonquest.betonquest.item.QuestItem;
-import org.betonquest.betonquest.modules.downloader.DownloadFailedException;
-import org.betonquest.betonquest.modules.downloader.Downloader;
 import org.betonquest.betonquest.modules.logger.BetonQuestLogRecord;
 import org.betonquest.betonquest.modules.logger.PlayerLogWatcher;
 import org.betonquest.betonquest.modules.logger.format.ChatFormatter;
 import org.betonquest.betonquest.modules.logger.handler.history.LogPublishingController;
-import org.betonquest.betonquest.modules.updater.Updater;
+import org.betonquest.betonquest.modules.web.downloader.DownloadFailedException;
+import org.betonquest.betonquest.modules.web.downloader.Downloader;
+import org.betonquest.betonquest.modules.web.updater.Updater;
 import org.betonquest.betonquest.utils.PlayerConverter;
 import org.betonquest.betonquest.utils.Utils;
 import org.bukkit.Bukkit;
