+++ src/main/java/org/betonquest/betonquest/modules/web/updater/UpdaterConfig.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.modules.updater;
+package org.betonquest.betonquest.modules.web.updater;
 
 import lombok.CustomLog;
 import org.betonquest.betonquest.modules.versioning.UpdateStrategy;
