+++ src/test/resources/modules/web/updater/github.json
