+++ src/test/resources/modules/web/updater/latest.json
