+++ src/test/resources/modules/web/updater/nexusPom1.xml
