+++ src/test/resources/modules/web/updater/nexusPom2.xml
